Propagate SourceMap into CircularReference (if available)
