+++ src/RefractAPI.cc
@@ -25,24 +25,6 @@ namespace drafter {
 
     namespace {
 
-        typedef std::vector<const snowcrash::DataStructure*> DataStructures;
-
-        void FindNamedTypes(const snowcrash::Elements& elements, DataStructures& found)
-        {
-            for (snowcrash::Elements::const_iterator i = elements.begin(); i != elements.end(); ++i) {
-
-                if (i->element == snowcrash::Element::DataStructureElement) {
-                    found.push_back(&(i->content.dataStructure));
-                }
-                else if (!i->content.resource.attributes.empty()) {
-                    found.push_back(&i->content.resource.attributes);
-                }
-                else if (i->element == snowcrash::Element::CategoryElement) {
-                    FindNamedTypes(i->content.elements(), found);
-                }
-            }
-        }
-
         template <typename T>
         bool IsNull(const T* ptr)
         {
@@ -59,7 +41,6 @@ namespace drafter {
         {
             NodeInfoCollection<T> nodeInfoCollection(nodeInfo);
             std::transform(nodeInfoCollection.begin(), nodeInfoCollection.end(), std::back_inserter(content), transformFunctor);
-
         }
 
         template<typename T, typename C, typename F>
@@ -81,171 +62,6 @@ namespace drafter {
             element->renderType(renderType);
 
             return element;
-        }
-    }
-
-    struct InheritanceComparator {
-
-        typedef std::map<std::string, std::string> InheritanceMap;
-        InheritanceMap childToParent;
-
-        typedef std::set<std::string> Members;
-        typedef std::map<std::string, Members> MembersMap;
-
-        MembersMap objectToMembers;
-
-        const std::string& parent(const snowcrash::DataStructure* ds) {
-            return ds->typeDefinition.typeSpecification.name.symbol.literal;
-        }
-
-        bool hasParent(const snowcrash::DataStructure* ds) {
-            return !parent(ds).empty();
-        }
-
-        const std::string& name(const snowcrash::DataStructure* ds) {
-            return ds->name.symbol.literal;
-        }
-
-        const std::string& name(const mson::ValueDefinition& vd) {
-            return vd.typeDefinition.typeSpecification.name.symbol.literal;
-        }
-
-        const std::string& name(const mson::ValueMember& vm) {
-            return name(vm.valueDefinition);
-        }
-
-        const std::string& name(const mson::PropertyMember& pm) {
-            return name(pm.valueDefinition);
-        }
-
-        Members collectMembers(const mson::Elements& elements) {
-            Members members;
-
-            for (mson::Elements::const_iterator it = elements.begin() ; it != elements.end() ; ++it) {
-
-                std::string member;
-                const mson::TypeSections* ts = NULL;
-
-                if (!it->content.value.empty() ) {
-                    if (!name(it->content.value).empty()) {
-                        member = name(it->content.value);
-                    }
-                    else {
-                        ts = &it->content.value.sections;
-                    }
-                }
-                else if (!it->content.property.empty()) {
-                    if(!name(it->content.property).empty()) {
-                        member = name(it->content.property);
-                    }
-                    else if (!name(it->content.property.name.variable).empty()) {
-                        member = name(it->content.property.name.variable);
-                    }
-                    else {
-                        ts = &it->content.property.sections;
-                    }
-                }
-
-                if (!member.empty()) {
-                    members.insert(member);
-                }
-                else if (ts) {
-                    Members sub = collectMembers(*ts);
-                    members.insert(sub.begin(), sub.end());
-                }
-            }
-
-            return members;
-        }
-
-        Members collectMembers(const mson::TypeSections& ts) {
-            Members members;
-
-            // map direct members
-            for (mson::TypeSections::const_iterator its = ts.begin() ; its != ts.end() ; ++its) {
-                Members sub = collectMembers(its->content.elements());
-                members.insert(sub.begin(), sub.end());
-            }
-
-            return members;
-        }
-
-        Members collectMembers(const snowcrash::DataStructure* ds) {
-            return collectMembers(ds->sections);
-        }
-
-        InheritanceComparator(const DataStructures& elements) {
-
-            // map inheritance
-            for (DataStructures::const_iterator i = elements.begin() ; i != elements.end() ; ++i) {
-                if (hasParent(*i)) {
-                    childToParent[name(*i)] = parent(*i);
-                }
-            }
-
-            for (DataStructures::const_iterator i = elements.begin() ; i != elements.end() ; ++i) {
-                objectToMembers[name(*i)] = collectMembers(*i);
-            }
-        }
-
-        bool hasMember(const snowcrash::DataStructure* object, const snowcrash::DataStructure* member) {
-            MembersMap::const_iterator members = objectToMembers.find(name(object));
-            if (members == objectToMembers.end()) {
-                return false;
-            }
-
-            return members->second.find(name(member)) != members->second.end();
-        }
-
-        bool hasAncestor(const snowcrash::DataStructure* object, const snowcrash::DataStructure* ancestor) {
-            std::string s = name(object);
-            const std::string& isAncestor = name(ancestor);
-
-            while(!s.empty()) {
-                if (s == isAncestor) {
-                    return true;
-                }
-                s = childToParent[s];
-            }
-
-            return false;
-        }
-
-
-        /* This is a comparator for std::sort so it has to compare
-         * objects in strictly weak ordering otherwise it would crash
-         * with sort going out of container bounds, see
-         * http://stackoverflow.com/questions/24048022/what-causes-stdsort-to-access-address-out-of-range
-         * for more details.
-         * The order is following, if F has S as an ancestor or a member
-         * then S < F and conversely, if none of this applies we compare objects by
-         * names to provide some ordering for all objects as we don't
-         * really care.
-         */
-        bool operator()(const snowcrash::DataStructure* first, const snowcrash::DataStructure* second) {
-            if (hasAncestor(first, second) || hasMember(first, second)) {
-                return false;
-            }
-            if (hasAncestor(second, first) || hasMember(second, first)) {
-                return true;
-            }
-            return name(first) < name (second);
-        }
-
-    };
-
-    void RegisterNamedTypes(const snowcrash::Elements& elements)
-    {
-        DataStructures found;
-        FindNamedTypes(elements, found);
-        std::sort(found.begin(), found.end(), InheritanceComparator(found));
-
-        for (DataStructures::const_iterator i = found.begin(); i != found.end(); ++i) {
-
-            if (!(*i)->name.symbol.literal.empty()) {
-                refract::IElement* element = MSONToRefract(MakeNodeInfoWithoutSourceMap(*(*i)), true);
-                GetNamedTypesRegistry().add(element);
-            }
         }
     }
